products/slack_bot/requirements.txt
slack-bolt>=1.18.0
Flask>=2.3.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
requests>=2.31.0